+++ src/components/PlayerBar.tsx
@@ -4,9 +4,8 @@ import type { SubsonicAlbum } from '../api/subsonicTypes';
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
 import {
-  Play, Pause, SkipBack, SkipForward, Volume2, VolumeX, Music,
-  Square, Repeat, Repeat1, Maximize2, SlidersVertical, X, Heart, Cast,
-  PictureInPicture2, ArrowLeftRight, Moon, Sunrise, Ellipsis,
+  SlidersVertical, X,
+  PictureInPicture2, Ellipsis,
 } from 'lucide-react';
 import { invoke } from '@tauri-apps/api/core';
 import { usePlayerStore } from '../store/playerStore';
@@ -34,6 +33,10 @@ import { PlaybackTime, RemainingTime } from './playerBar/PlaybackClock';
 import { PlayerTrackInfo } from './playerBar/PlayerTrackInfo';
 import { PlayerTransportControls } from './playerBar/PlayerTransportControls';
 import { PlayerSeekbarSection } from './playerBar/PlayerSeekbarSection';
+import { PlayerVolume } from './playerBar/PlayerVolume';
+import { PlayerOverflowMenu } from './playerBar/PlayerOverflowMenu';
+import { useFloatingPlayerBar } from '../hooks/useFloatingPlayerBar';
+import { useUtilityOverflowMenu } from '../hooks/useUtilityOverflowMenu';
 
 export default function PlayerBar() {
   const { t } = useTranslation();
@@ -79,137 +82,30 @@ export default function PlayerBar() {
   })));
   const { lastfmSessionKey } = useAuthStore();
   const floatingPlayerBar = useThemeStore(s => s.floatingPlayerBar);
-  const [floatingStyle, setFloatingStyle] = useState<React.CSSProperties>({});
   const playerBarRef = useRef<HTMLElement>(null);
-  const [utilityOverflow, setUtilityOverflow] = useState(false);
-  const [utilityMenuOpen, setUtilityMenuOpen] = useState(false);
-  const [utilityMenuMode, setUtilityMenuMode] = useState<'full' | 'volume'>('full');
-  const utilityMenuRef = useRef<HTMLDivElement>(null);
-  const utilityBtnRef = useRef<HTMLButtonElement>(null);
-  const [utilityMenuStyle, setUtilityMenuStyle] = useState<React.CSSProperties>({});
-  const volumeWheelMenuTimerRef = useRef<number | null>(null);
-  const [suppressOverflowTooltip, setSuppressOverflowTooltip] = useState(false);
   const perfFlags = usePerfProbeFlags();
 
-  useEffect(() => {
-    if (!floatingPlayerBar) return;
+  const floatingStyle = useFloatingPlayerBar(playerBarRef, floatingPlayerBar);
 
-    const updatePosition = () => {
-      const sidebar = document.querySelector('.sidebar') as HTMLElement;
-      const queue = document.querySelector('.queue-panel') as HTMLElement;
-
-      const leftOffset = sidebar ? sidebar.getBoundingClientRect().right : 0;
-      const rightOffset = queue ? window.innerWidth - queue.getBoundingClientRect().left : 0;
-
-      setFloatingStyle({
-        left: leftOffset + 24,
-        right: rightOffset + 24,
-        width: 'auto',
-      });
-    };
-
-    updatePosition();
-
-    const observer = new ResizeObserver(updatePosition);
-    const sidebar = document.querySelector('.sidebar');
-    const queue = document.querySelector('.queue-panel');
-    if (sidebar) observer.observe(sidebar);
-    if (queue) observer.observe(queue);
-    window.addEventListener('resize', updatePosition);
-
-    return () => {
-      observer.disconnect();
-      window.removeEventListener('resize', updatePosition);
-    };
-  }, [floatingPlayerBar]);
-
-  useEffect(() => {
-    const updateOverflow = () => {
-      const width = playerBarRef.current?.clientWidth ?? window.innerWidth;
-      const threshold = floatingPlayerBar ? 980 : 1140;
-      setUtilityOverflow(width < threshold);
-    };
-
-    updateOverflow();
-    const ro = typeof ResizeObserver !== 'undefined'
-      ? new ResizeObserver(updateOverflow)
-      : null;
-    const el = playerBarRef.current;
-    if (ro && el) ro.observe(el);
-    window.addEventListener('resize', updateOverflow);
-    return () => {
-      ro?.disconnect();
-      window.removeEventListener('resize', updateOverflow);
-    };
-  }, [floatingPlayerBar]);
-
-  useEffect(() => {
-    if (!utilityOverflow) setUtilityMenuOpen(false);
-    if (!utilityOverflow && volumeWheelMenuTimerRef.current != null) {
-      window.clearTimeout(volumeWheelMenuTimerRef.current);
-      volumeWheelMenuTimerRef.current = null;
-    }
-  }, [utilityOverflow]);
-
-  useEffect(() => {
-    if (!utilityMenuOpen) return;
-    const onDown = (e: MouseEvent) => {
-      const target = e.target as Node;
-      if (utilityBtnRef.current?.contains(target)) return;
-      if (utilityMenuRef.current?.contains(target)) return;
-      setUtilityMenuOpen(false);
-    };
-    const onKey = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') setUtilityMenuOpen(false);
-    };
-    document.addEventListener('mousedown', onDown);
-    document.addEventListener('keydown', onKey);
-    return () => {
-      document.removeEventListener('mousedown', onDown);
-      document.removeEventListener('keydown', onKey);
-    };
-  }, [utilityMenuOpen]);
-
-  useEffect(() => () => {
-    if (volumeWheelMenuTimerRef.current != null) {
-      window.clearTimeout(volumeWheelMenuTimerRef.current);
-    }
-  }, []);
+  const {
+    utilityOverflow,
+    utilityMenuOpen,
+    setUtilityMenuOpen,
+    utilityMenuMode,
+    setUtilityMenuMode,
+    utilityMenuStyle,
+    utilityMenuRef,
+    utilityBtnRef,
+    volumeWheelMenuTimerRef,
+    suppressOverflowTooltip,
+    setSuppressOverflowTooltip,
+  } = useUtilityOverflowMenu(playerBarRef, floatingPlayerBar);
 
   useEffect(() => {
     const onToggleEqualizer = () => setEqOpen(v => !v);
     window.addEventListener('psy:toggle-equalizer', onToggleEqualizer);
     return () => window.removeEventListener('psy:toggle-equalizer', onToggleEqualizer);
   }, []);
-
-  useEffect(() => {
-    if (!utilityMenuOpen) return;
-    const MENU_WIDTH = 238;
-    const MARGIN = 8;
-    const updateMenuPos = () => {
-      const btn = utilityBtnRef.current;
-      if (!btn) return;
-      const r = btn.getBoundingClientRect();
-      const left = Math.min(
-        Math.max(r.right - MENU_WIDTH, MARGIN),
-        window.innerWidth - MENU_WIDTH - MARGIN,
-      );
-      setUtilityMenuStyle({
-        position: 'fixed',
-        left,
-        width: MENU_WIDTH,
-        bottom: window.innerHeight - r.top + 8,
-        zIndex: 10050,
-      });
-    };
-    updateMenuPos();
-    window.addEventListener('resize', updateMenuPos);
-    window.addEventListener('scroll', updateMenuPos, true);
-    return () => {
-      window.removeEventListener('resize', updateMenuPos);
-      window.removeEventListener('scroll', updateMenuPos, true);
-    };
-  }, [utilityMenuOpen]);
 
   const { delayModalOpen, setDelayModalOpen, playPauseBind } = usePlaybackDelayPress(togglePlay);
   const transportAnchorRef = useRef<HTMLDivElement>(null);
@@ -375,7 +271,6 @@ export default function PlayerBar() {
         </div>
       ) : (
         <>
-          {/* EQ Button */}
           <button
             className={`player-btn player-btn-sm player-eq-btn ${eqOpen ? 'active' : ''}`}
             onClick={() => setEqOpen(v => !v)}
@@ -385,7 +280,6 @@ export default function PlayerBar() {
             <SlidersVertical size={15} />
           </button>
 
-          {/* Mini Player */}
           <button
             className="player-btn player-btn-sm"
             onClick={() => invoke('open_mini_player').catch(() => {})}
@@ -395,161 +289,39 @@ export default function PlayerBar() {
             <PictureInPicture2 size={15} />
           </button>
 
-          {/* Volume */}
-          <div className="player-volume-section">
-            <button
-              className="player-btn player-btn-sm"
-              onClick={() => {
-                if (volume === 0) {
-                  setVolume(premuteVolumeRef.current);
-                } else {
-                  premuteVolumeRef.current = volume;
-                  setVolume(0);
-                }
-              }}
-              aria-label={t('player.volume')}
-              style={{ color: 'var(--text-muted)', flexShrink: 0 }}
-            >
-              {volume === 0 ? <VolumeX size={16} /> : <Volume2 size={16} />}
-            </button>
-            <div className="player-volume-slider-wrap" onWheel={handleVolumeWheel}>
-              {showVolPct && (
-                <span className="player-volume-pct" style={{ left: `${volume * 100}%` }}>
-                  {Math.round(volume * 100)}%
-                </span>
-              )}
-              <input
-                type="range"
-                id="player-volume"
-                min={0}
-                max={1}
-                step={0.01}
-                value={volume}
-                onChange={handleVolume}
-                style={volumeStyle}
-                aria-label={t('player.volume')}
-                className="player-volume-slider"
-                onMouseEnter={() => setShowVolPct(true)}
-                onMouseLeave={() => setShowVolPct(false)}
-              />
-            </div>
-          </div>
+          <PlayerVolume
+            volume={volume}
+            setVolume={setVolume}
+            premuteVolumeRef={premuteVolumeRef}
+            showVolPct={showVolPct}
+            setShowVolPct={setShowVolPct}
+            handleVolume={handleVolume}
+            handleVolumeWheel={handleVolumeWheel}
+            volumeStyle={volumeStyle}
+            inputId="player-volume"
+            t={t}
+          />
         </>
       )}
 
-      {/* EQ Popup — rendered via portal to avoid backdrop-filter containing-block issue */}
-      {utilityMenuOpen && createPortal(
-        <div
-          className={`player-overflow-menu${utilityMenuMode === 'volume' ? ' player-overflow-menu--volume-only' : ''}`}
-          ref={utilityMenuRef}
-          style={utilityMenuStyle}
-          onWheel={handleVolumeWheel}
-        >
-          {utilityMenuMode === 'full' && (
-            <div className="player-overflow-menu-row">
-              <button
-                className={`player-overflow-menu-btn${eqOpen ? ' active' : ''}`}
-                onClick={() => {
-                  setEqOpen(v => !v);
-                  setUtilityMenuOpen(false);
-                }}
-              >
-                <SlidersVertical size={14} />
-                {t('player.equalizer')}
-              </button>
-              <button
-                className="player-overflow-menu-btn"
-                onClick={() => {
-                  invoke('open_mini_player').catch(() => {});
-                  setUtilityMenuOpen(false);
-                }}
-              >
-                <PictureInPicture2 size={14} />
-                {t('player.miniPlayer')}
-              </button>
-            </div>
-          )}
-          {utilityMenuMode === 'full' ? (
-            <div className="player-volume-section player-volume-section--menu">
-              <button
-                className="player-btn player-btn-sm"
-                onClick={() => {
-                  if (volume === 0) {
-                    setVolume(premuteVolumeRef.current);
-                  } else {
-                    premuteVolumeRef.current = volume;
-                    setVolume(0);
-                  }
-                }}
-                aria-label={t('player.volume')}
-                style={{ color: 'var(--text-muted)', flexShrink: 0 }}
-              >
-                {volume === 0 ? <VolumeX size={16} /> : <Volume2 size={16} />}
-              </button>
-              <div className="player-volume-slider-wrap" onWheel={handleVolumeWheel}>
-                {showVolPct && (
-                  <span className="player-volume-pct" style={{ left: `${volume * 100}%` }}>
-                    {Math.round(volume * 100)}%
-                  </span>
-                )}
-                <input
-                  type="range"
-                  id="player-volume-overflow"
-                  min={0}
-                  max={1}
-                  step={0.01}
-                  value={volume}
-                  onChange={handleVolume}
-                  style={volumeStyle}
-                  aria-label={t('player.volume')}
-                  className="player-volume-slider"
-                  onMouseEnter={() => setShowVolPct(true)}
-                  onMouseLeave={() => setShowVolPct(false)}
-                />
-              </div>
-            </div>
-          ) : (
-            <div className="player-volume-section player-volume-section--menu">
-              <button
-                className="player-btn player-btn-sm"
-                onClick={() => {
-                  if (volume === 0) {
-                    setVolume(premuteVolumeRef.current);
-                  } else {
-                    premuteVolumeRef.current = volume;
-                    setVolume(0);
-                  }
-                }}
-                aria-label={t('player.volume')}
-                style={{ color: 'var(--text-muted)', flexShrink: 0 }}
-              >
-                {volume === 0 ? <VolumeX size={16} /> : <Volume2 size={16} />}
-              </button>
-              <div className="player-volume-slider-wrap player-volume-slider-wrap--menu-only" onWheel={handleVolumeWheel}>
-                {showVolPct && (
-                  <span className="player-volume-pct" style={{ left: `${volume * 100}%` }}>
-                    {Math.round(volume * 100)}%
-                  </span>
-                )}
-                <input
-                  type="range"
-                  id="player-volume-overflow-wheel"
-                  min={0}
-                  max={1}
-                  step={0.01}
-                  value={volume}
-                  onChange={handleVolume}
-                  style={volumeStyle}
-                  aria-label={t('player.volume')}
-                  className="player-volume-slider"
-                  onMouseEnter={() => setShowVolPct(true)}
-                  onMouseLeave={() => setShowVolPct(false)}
-                />
-              </div>
-            </div>
-          )}
-        </div>,
-        document.body
+      {utilityMenuOpen && (
+        <PlayerOverflowMenu
+          utilityMenuRef={utilityMenuRef}
+          utilityMenuStyle={utilityMenuStyle}
+          utilityMenuMode={utilityMenuMode}
+          eqOpen={eqOpen}
+          setEqOpen={setEqOpen}
+          closeMenu={() => setUtilityMenuOpen(false)}
+          volume={volume}
+          setVolume={setVolume}
+          premuteVolumeRef={premuteVolumeRef}
+          showVolPct={showVolPct}
+          setShowVolPct={setShowVolPct}
+          handleVolume={handleVolume}
+          handleVolumeWheel={handleVolumeWheel}
+          volumeStyle={volumeStyle}
+          t={t}
+        />
       )}
 
       {/* EQ Popup — rendered via portal to avoid backdrop-filter containing-block issue */}
